% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qc_on_table.R
\name{check_table_name}
\alias{check_table_name}
\title{Check that protocol table is defined in schema}
\usage{
check_table_name(protocol, table)
}
\arguments{
\item{table}{}
}
\description{
Check that protocol table is defined in schema
}
