% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_column_evaluation.R
\name{column_evaluation}
\alias{column_evaluation}
\title{Column Evaluation Test}
\usage{
column_evaluation(df, column_order_list)
}
\arguments{
\item{column_order_list}{}
}
\description{
Column Evaluation Test
}
