% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_coordinates.R
\name{coordinates_present}
\alias{coordinates_present}
\title{Check if coordinates are defined in sample event transect metadata table}
\usage{
coordinates_present(data_list, p)
}
\arguments{
\item{p}{}
}
\description{
Check if coordinates are defined in sample event transect metadata table
}
