% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qc_on_table.R
\name{create_quality_control_summary}
\alias{create_quality_control_summary}
\title{Create Quality Control Summary}
\usage{
create_quality_control_summary(quality_control_results)
}
\arguments{
\item{quality_control_results}{}
}
\description{
Create Quality Control Summary
}
