% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_schema.R
\name{create_schema_list}
\alias{create_schema_list}
\title{Create schema list}
\usage{
create_schema_list(filepath_to_data_structure)
}
\arguments{
\item{filepath_to_data_structure}{}
}
\description{
Create schema list
}
