% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_tables.R
\name{get_column_order}
\alias{get_column_order}
\title{Get Column Order for table}
\usage{
get_column_order(protocol, table, dataset = "L2_standardized")
}
\arguments{
\item{level}{}
}
\description{
Get Column Order for table
}
