% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_event_from_directory.R
\name{load_sample_event_data}
\alias{load_sample_event_data}
\title{Load Sample Event from directory}
\usage{
load_sample_event_data(directory)
}
\arguments{
\item{directory}{}
}
\description{
Load Sample Event from directory
}
