% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qc_on_sample_event.R
\name{run_qc_on_sample_event}
\alias{run_qc_on_sample_event}
\title{Run QC on sample event}
\usage{
run_qc_on_sample_event(data_list, dataset = "L2_standardized")
}
\arguments{
\item{data_list}{}
}
\description{
Run QC on sample event
}
