% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qc_on_sample_event.R
\name{run_sample_event_summary}
\alias{run_sample_event_summary}
\title{Conduct sample event summary}
\usage{
run_sample_event_summary(data_list, grouping)
}
\arguments{
\item{grouping}{}
}
\description{
Conduct sample event summary
}
