% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sample_event_summary.R
\name{sample_event_summary}
\alias{sample_event_summary}
\title{Summarize sample event}
\usage{
sample_event_summary(data_list, p)
}
\arguments{
\item{p}{}
}
\description{
Summarize sample event
}
