% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qc_on_table.R
\name{table_test_controller}
\alias{table_test_controller}
\title{Run QC tests on tables}
\usage{
table_test_controller(df, protocol, table)
}
\arguments{
\item{table}{}
}
\description{
Run QC tests on tables
}
