% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_tables.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Check order and missing or invalid columns}
\usage{
validate_data_structure(df, column_order)
}
\arguments{
\item{column_order}{}
}
\description{
Check order and missing or invalid columns
}
