% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_data_directory.R
\name{write_data_directory_to_L3}
\alias{write_data_directory_to_L3}
\title{Title}
\usage{
write_data_directory_to_L3(destination, file_inventory, sample_event_id)
}
\arguments{
\item{sample_event_id}{}
}
\description{
Title
}
