% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_schema.R
\name{write_schema_json}
\alias{write_schema_json}
\title{Write Schema JSON}
\usage{
write_schema_json(schema, knowledge_hub_list)
}
\arguments{
\item{knowledge_hub_list}{}
}
\description{
Write Schema JSON
}
